Enable scrollbar with commands, not an option.
An option is a nice UI because it feels more "built in", but it introduces a
load-order dependency - only plugins loaded *after* the scrollbar plugin can set
the scrollbar option.

Using commands to enable and disable the scrollbar means they can be called from
hooks, which won't be invoked until after all plugins are loaded, solving the
load-order dependency. In addition, this makes things a little easier to debug
when things go wrong - if the user puts a hook in their `kakrc`, they know they
have to debug hooks; if they set an option in their `kakrc` they may spend time
debugging options when options aren't the problem. For precedent, see the
`autowrap-enable` and `ctags-enable-*` commands in the standard library.

Also, I took the liberty of inlining some helper functions. The "enable" and
"disable" logic can now be viewed side-by-side to verify that everything that
gets installed also gets cleaned up. Added comments mean it's still possible to
understand what the code does without being a Kakoune expert.