Calculate bar height based on window_range, not window_height.
window_range describes the actual number of visible lines, which determines how
many flags we can actually place.

window_height is the number of lines that could *theoretically* be shown, if
there were no soft-wrapped lines.

Of the two, window_range is the closest match for the purposes we want.